requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv

PyPDF2
python-docx

langchain
langchain-google-genai

